requirements.txt
playwright
pytest
allure-pytest
pytest-base-url
pytest-playwright
tempmail-python
faker